backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
httpx==0.26.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic[email]==2.5.3
deepgram-sdk==3.4.0
websockets==12.0
google-generativeai==0.8.5
groq==0.9.0
